Add new features to the HYPERWORLD Datastore website
- Update get.js to fetch and display game data
- Modify index.html to improve styling and layout
- Fix minor bugs in existing code

This commit adds new functionality to the HYPERWORLD Datastore website by updating the get.js file to fetch and display game data. It also improves the styling and layout of the website in index.html and fixes minor bugs in existing code. This will make the website more user-friendly and provide more accurate game information for users.